pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.2",
    "channels>=4.2.2",
    "channels-redis>=4.2.1",
    "crispy-bootstrap5>=2025.4",
    "daphne>=4.1.2",
    "dj-database-url>=2.3.0",
    "django>=5.2",
    "django-crispy-forms>=2.4",
    "django-filter>=25.1",
    "gunicorn>=23.0.0",
    "openpyxl>=3.1.5",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "redis>=6.0.0",
    "reportlab>=4.4.0",
]
